fix: Move resource result types into 1 file
