+++ dhis-2/dhis-api/src/main/java/org/hisp/dhis/appmanager/AppManager.java
@@ -34,7 +34,6 @@ import java.util.List;
 import java.util.Set;
 import java.util.UUID;
 import java.util.stream.Collectors;
-import org.hisp.dhis.appmanager.resource.ResourceResult;
 import org.springframework.core.io.Resource;
 
 /**
