+++ dhis-2/dhis-test-integration/src/test/java/org/hisp/dhis/app/AppManagerMinIOTest.java
@@ -38,9 +38,9 @@ import java.util.stream.Stream;
 import org.hisp.dhis.appmanager.App;
 import org.hisp.dhis.appmanager.AppManager;
 import org.hisp.dhis.appmanager.AppStatus;
-import org.hisp.dhis.appmanager.resource.Redirect;
-import org.hisp.dhis.appmanager.resource.ResourceFound;
-import org.hisp.dhis.appmanager.resource.ResourceNotFound;
+import org.hisp.dhis.appmanager.ResourceResult.Redirect;
+import org.hisp.dhis.appmanager.ResourceResult.ResourceFound;
+import org.hisp.dhis.appmanager.ResourceResult.ResourceNotFound;
 import org.hisp.dhis.test.integration.PostgresIntegrationTestBase;
 import org.hisp.dhis.test.junit.MinIOTestExtension;
 import org.hisp.dhis.test.junit.MinIOTestExtension.DhisConfig;
